feat: 산책 UI 개선 및 컴포넌트 리팩토링
왕복 시간 뷰 추가 및 산책로 아이템에 시간, 아이콘 기능 통합
관련 홈뷰 컴포넌트 명칭 변경과 레이아웃 간격 조정
